requirements.txt
krakenex==2.1.0
pandas==0.25.1
APScheduler==3.7.0
psycopg2-binary==2.8.6
